+++ core/src/com/bnana/goa/utils/ScaleManager.java
@@ -4,5 +4,5 @@ package com.bnana.goa.utils;
  * Created by luca.piccinelli on 21/09/2015.
  */
 public interface ScaleManager {
-    float s(int n);
+    float s(float n);
 }
